finished styling home page
