+++ client/src/pages/Home.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from "react";
+import { Link } from 'react-router-dom'
 import styles from '../styles/Home.module.css';
 
 export function Home() {
@@ -13,6 +14,19 @@ export function Home() {
     return () => clearInterval(intervalId);
   }, []);
 
+  async function logout() {
+    const res = await fetch("/registration/logout/", {
+      credentials: "same-origin", // include cookies!
+    });
+
+    if (res.ok) {
+      // navigate away from the single page app!
+      window.location = "/registration/sign_in/";
+    } else {
+      console.error("This was a stupid request")
+    }
+  }
+
   return (
     <>
       <nav className={styles.navbar}>
@@ -20,16 +34,24 @@ export function Home() {
           <span className={styles.navbarTitle}>Meal Deal</span>
         </div>
         <div className={styles.navbarRight}>
-          <button className={styles.navbarButton}>View Plans</button>
-          <button className={styles.navbarButton}>Create Plans</button>
-          <button className={styles.navbarButton}>Logout</button>
+            <Link to="/view_plans">
+                <button className={styles.navbarButton}>View Plans</button>
+            </Link>
+            <Link to="/create_plan">
+                <button className={styles.navbarButton}>Create Plans</button>
+            </Link>
+          <button className={styles.navbarButton} onClick={logout}>Logout</button>
         </div>
       </nav>
       <section className={styles.container}>
         <div className={styles.info}>
             <h1 className={styles.helloUser}>Hi, User!</h1>
-            <p className={styles.date}>Current Date: {dateTime.toLocaleDateString()}</p>
-
+            <p className={styles.date}>
+                Today's Date: {dateTime.toLocaleDateString('en-US', { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' })}
+            </p>
+            <div className={styles.meal}>
+                This is where today's meals will go
+            </div>
         </div>
       </section>
     </>
